requirements.txt
praw
discord.py==1.7.2
python-dotenv